Move firstprograms to src root
