--- firstprograms/SimpleEx.java
@@ -1,28 +1,0 @@
-package com.zetcode;
-
-import java.awt.EventQueue;
-import javax.swing.JFrame;
-
-public class SimpleEx extends JFrame {
-
-    public SimpleEx() {
-
-        initUI();
-    }
-
-    private void initUI() {
-        
-        setTitle("Simple example");
-        setSize(300, 200);
-        setLocationRelativeTo(null);
-        setDefaultCloseOperation(EXIT_ON_CLOSE);
-    }
-
-    public static void main(String[] args) {
-
-        EventQueue.invokeLater(() -> {
-            var ex = new SimpleEx();
-            ex.setVisible(true);
-        });
-    }
-}
