+++ lib/src/main/java/com/zetcode/firstprograms/TooltipEx.java
@@ -1,54 +1,58 @@
-package com.zetcode;
+package com.zetcode.firstprograms;
 
+import java.awt.EventQueue;
 import javax.swing.GroupLayout;
 import javax.swing.JButton;
 import javax.swing.JComponent;
 import javax.swing.JFrame;
-import java.awt.EventQueue;
+import javax.swing.JPanel;
 
-public class QuitButtonEx extends JFrame {
+public class TooltipEx extends JFrame {
 
-    public QuitButtonEx() {
+    public TooltipEx() {
 
         initUI();
     }
 
     private void initUI() {
 
-        var quitButton = new JButton("Quit");
+        var btn = new JButton("Button");
+        btn.setToolTipText("A button component");
 
-        quitButton.addActionListener(_ -> System.exit(0));
+        createLayout(btn);
 
-        createLayout(quitButton);
-
-        setTitle("Quit button");
-        setSize(300, 200);
+        setTitle("Tooltip");
         setLocationRelativeTo(null);
         setDefaultCloseOperation(EXIT_ON_CLOSE);
     }
 
     private void createLayout(JComponent... arg) {
 
-        var pane = getContentPane();
+        var pane = (JPanel) getContentPane();
         var gl = new GroupLayout(pane);
         pane.setLayout(gl);
+
+        pane.setToolTipText("Content pane");
 
         gl.setAutoCreateContainerGaps(true);
 
         gl.setHorizontalGroup(gl.createSequentialGroup()
                 .addComponent(arg[0])
+                .addGap(200)
         );
 
         gl.setVerticalGroup(gl.createSequentialGroup()
                 .addComponent(arg[0])
+                .addGap(120)
         );
+
+        pack();
     }
 
     public static void main(String[] args) {
 
         EventQueue.invokeLater(() -> {
-            
-            var ex = new QuitButtonEx();
+            var ex = new TooltipEx();
             ex.setVisible(true);
         });
     }
